[packages/template][s]: move e2e tests to separate folder
