--- packages/template-e2e/cypress.config.js
@@ -1,5 +1,0 @@
-import { defineConfig } from "cypress";
-import { nxE2EPreset } from "@nrwl/cypress/plugins/cypress-preset";
-export default defineConfig({
-  e2e: nxE2EPreset(__dirname),
-});
