--- packages/template-e2e/src/e2e/app.cy.js
@@ -1,10 +1,0 @@
-import { getGreeting } from "../support/app.po";
-describe("template", () => {
-  beforeEach(() => cy.visit("/"));
-  it("should display welcome message", () => {
-    // Custom command example, see `../support/commands.ts` file
-    cy.login("my-email@something.com", "myPassword");
-    // Function helper example, see `../support/app.po.ts` file
-    getGreeting().contains("Welcome template");
-  });
-});
