+++ packages/template-e2e/src/fixtures/content/assets/images/parktest.png
