+++ packages/template-e2e/src/fixtures/content/config.js
@@ -1,0 +1,1 @@
+export default {};
