--- packages/template-e2e/src/support/app.po.js
@@ -1,1 +1,0 @@
-export const getGreeting = () => cy.get("h1");
