+++ packages/template-e2e/src/support/testsetup.sh
@@ -1,0 +1,18 @@
+#!/usr/bin/sh
+
+TEMPLATE_DIR=$1
+CONTENT_DIR=$2
+E2E_TEMP_DIR=$(mktemp -d)
+
+cp -r $TEMPLATE_DIR $E2E_TEMP_DIR
+cp -r $CONTENT_DIR $E2E_TEMP_DIR
+
+TEMP_TEMPLATE_DIR=$E2E_TEMP_DIR/$(basename $TEMPLATE_DIR)
+TEMP_CONTENT_DIR=$E2E_TEMP_DIR/$(basename $CONTENT_DIR)
+
+ln -vfns $TEMP_CONTENT_DIR $TEMP_TEMPLATE_DIR/content
+ln -vfns $TEMP_CONTENT_DIR/assets $TEMP_TEMPLATE_DIR/public/assets
+
+cd $TEMP_TEMPLATE_DIR
+npm install
+npm run dev &
