+++ packages/template-e2e/src/tests/pages.spec.js
