--- packages/template/testsetup.sh
@@ -1,4 +1,0 @@
-rm content
-ln -s tests/fixtures/content content
-rm public/assets
-ln -s ../tests/fixtures/content/assets public/assets
